all-in on immutable + better crate comment
